!HotFix: CountBy 반환값 Long 변경 CountBy 는 기본적으로 Long 타입을 반환합니다. 따라서 Integer 타입으로 설정 시 예측되지 않는 데이터를 반환합니다.

+++ src/main/java/com/knu/ntttt_server/token/controller/GachaTokenController.java
@@ -34,7 +34,7 @@ public class GachaTokenController {
 
     @Operation(summary = "같은 토큰을 뽑은 유저의 수를 조회.", description = "같은 토큰을 뽑은 유저의 수를 조회합니다.")
     @GetMapping("/{tokenId}")
-    public ApiResponse<Integer> getUserNumberSeeingSameToken(@PathVariable Long tokenId) {
+    public ApiResponse<Long> getUserNumberSeeingSameToken(@PathVariable Long tokenId) {
         return ApiResponse.ok(gachaService.getUserNumberSeeingSameToken(tokenId));
     }
 }
